HostelFix/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
cloudinary==1.38.0
pydantic==2.5.0
pydantic-settings==2.1.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic-settings==2.1.0

gunicorn==21.2.0


email-validator==2.1.0

numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
